Some decent changes to types...
1.  Alias `#===` to test so types work in case statements

2.  Introduce {NRSER::Types::When} that just wraps an object and uses
`#===` as it's test; default to it in {NRSER::Types.make} (when not
already a type) instead of the messy mess of IsA and Is it was before.

    This makes a *lot* more sense. May switch everything over to `#===`
and not wrap these objects in the future, but this is enough shift for
one day. Much better tho!

3.  Finally add {NRSER::Types::Shape}... it's not much, but it's a
start.

4.  Uh a buncha twiddle-fiddle with type printing... it looks all mathy
kind-of. We'll see how it feel after a few.

5.  Etc. :)