Experimental: Have {NRSER::NicerError} look for `#to_summary`
For a while I've wanted a "summary" method that can be added to classes
to provide short summaries of their contents suitable for inline use...
this is a first step in that direction.

This idea kind-of existed in RSpex as `#desc` and `#to_short_s`, but -
at the moment, at least - I think "summary" is a better name for it.